requirements.txt
Django==1.8.0
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
djangorestframework==3.0.0
gunicorn==19.1.1
PyMySQL==0.7.1
mysqlclient==1.3.7
psycopg2==2.5.4
static3==0.6.1
django-user-accounts==2.0.0
pinax-theme-bootstrap==7.11.0
